feta: LoginCoordinator
